Implements windows memory stats

+++ src/eatmemory.apple.c
@@ -5,8 +5,6 @@
     #include <mach/mach_init.h>
 
     void get_system_memory_stats(struct system_memory_stats* stats) {
-
-
         vm_size_t page_size;
         mach_port_t host_port = mach_host_self();
         mach_msg_type_number_t count;
@@ -23,6 +21,7 @@
 
         natural_t pages = vm_stat.wire_count + vm_stat.active_count + vm_stat.inactive_count + vm_stat.free_count +
                           vm_stat.speculative_count + vm_stat.purgeable_count;
+        
         stats->total = (size_t)pages * (size_t)page_size;
 
         natural_t free_memory = vm_stat.free_count + vm_stat.inactive_count;
@@ -42,6 +41,6 @@
         }
         
         stats->supported = true;
-        stats->rss = task_basic_info.resident_size;
+        stats->rss = (size_t)task_basic_info.resident_size;
     }
 #endif
