+++ src/eatmemory.linux.c
@@ -8,10 +8,11 @@ void get_system_memory_stats(struct system_memory_stats* stats) {
     stats->supported = true;
     long pages = sysconf(_SC_PHYS_PAGES);
     long page_size = sysconf(_SC_PAGE_SIZE);
-    stats->total = pages * page_size;
+    
+    stats->total = (size_t)pages * (size_t)page_size;
 
     long free_pages = sysconf(_SC_AVPHYS_PAGES);
-    stats->free = free_pages * page_size;
+    stats->free = (size_t)free_pages * (size_t)page_size;
 }
 
 void get_process_memory_stats(struct process_memory_stats* stats) {
@@ -26,9 +27,9 @@ void get_process_memory_stats(struct process_memory_stats* stats) {
     while (fgets(line, sizeof(line), file)) {
         if (strncmp(line, "VmRSS:", 6) == 0) {
             // Parse VmRSS value in kB
-            size_t rss_kb;
-            if (sscanf(line + 6, "%zu", &rss_kb) == 1) {
-                stats->rss = rss_kb * 1024; // Convert from kB to bytes
+            unsigned long long rss_kb;
+            if (sscanf(line + 6, "%llu", &rss_kb) == 1) {
+                stats->rss = (size_t)(rss_kb * 1024); // Convert from kB to bytes
                 stats->supported = true;
             }
             break;
